+++ DataModels/LinkType.cs
@@ -1,0 +1,8 @@
+namespace VidsNet.Enums
+{
+    public enum LinkType {
+        Both,
+        Desktop,
+        Mobile
+    }
+}