+++ Migrations/20160826095611_SystemMessage_LongMessage.cs
@@ -1,0 +1,24 @@
+﻿using System;
+using System.Collections.Generic;
+using Microsoft.EntityFrameworkCore.Migrations;
+
+namespace vidsnet.Migrations
+{
+    public partial class SystemMessage_LongMessage : Migration
+    {
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "LongMessage",
+                table: "SystemMessages",
+                nullable: true);
+        }
+
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "LongMessage",
+                table: "SystemMessages");
+        }
+    }
+}