+++ public/javascripts/scripts.js
@@ -13,9 +13,7 @@ function updateVirtualView() {
 function changeVirtualView() {
     var cb = function(xhr) {
         if(xhr.status == 200) {
-            var txt = xhr.responseText.replace(new RegExp("&lt;", 'g'), "<").replace(new RegExp("&gt;", 'g'), ">");
-            
-            $("#contentBox").html(txt);
+            updateContent(xhr.responseText);
             updatePopovers();
         }
     };
@@ -54,6 +52,37 @@ $( document ).ready(function() {
     updatePopovers();
 });
 
-function openSystemMessages() {
-    console.log("System messages are now opened!");
+function openMessage(id) {
+    var cb = function(xhr) {
+        if(xhr.status === 200) {
+            openModal("System message", xhr.responseText, true);
+        }
+    };
+
+    sendQuery("/systemmessages/get/" + id, null, "GET", cb);
+}
+
+function cleanMessages() {
+    var cb = function(xhr) {
+        if(xhr.status === 200) {
+            var cb2 = function(xhr2) {
+                if(xhr2.status === 200) {
+                    updateContent(xhr2.responseText);
+                }
+            };
+            sendQuery("/systemmessages/getall", null, "GET", cb2);
+        }
+    };
+
+    sendQuery("/systemmessages/clean", null, "GET", cb);
+}
+
+function deleteMessage(id) {
+    var cb = function(xhr) {
+        if(xhr.status === 200) {
+            setHide(id + "_tr");
+        }
+    };
+
+    sendQuery("/systemmessages/delete/" + id, null, "DELETE", cb);
 }